Created a HeroAI that will read AIBehaviours
Went back to the enum idea and create small functions inside it, but uh now the gates won't open! This is due to the nature of how they read values and accept values considering I have now a keyIndex inside AIBehaviour and SmartHero I should make it READ that instead of GameVariables! But... I could use that for a external reference...?